requirements.txt
requests>=2.31.0
websocket-client>=1.7.0
pyinstaller>=6.0.0
